requirements.txt
pytube
requests
discord.py
pynacl